pyproject.toml
[project]
name = "ocs-ci"
version = "4.21.0"
description = "ODF QE test framework for qualifying Data Foundation"
authors = [{name = "OCS QE", email = "ocs-ci@redhat.com"}]
license="MIT"
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
    "apache-libcloud==3.8.0",
    "atlassian-python-api==4.0.7",
    "azure-common==1.1.28",
    "azure-identity==1.16.1",
    "azure-keyvault-secrets==4.8.0",
    "azure-mgmt-compute==33.0.0",
    "azure-mgmt-network==28.0.0",
    "azure-mgmt-resource==23.2.0",
    "azure-mgmt-storage==21.0.0",
    "azure-storage-blob==12.23.1",
    "boto3==1.38.31",
    "bs4==0.0.1",
    "certbot==3.0.0",
    "certbot-dns-route53==3.0.0",
    "docopt==0.6.2",
    "elasticsearch==8.11.1",
    "fauxfactory==3.1.0",
    "flaky==3.7.0",
    "funcy==1.14",
    "gevent==23.9.1",
    "gitpython==3.1.41",
    "google-api-core==2.25.0",
    "google-api-python-client==2.171.0",
    "google-auth==2.38.0",  # pragma: allowlist secret
    "google-auth-httplib2==0.2.0",
    "google-auth-oauthlib==1.2.1",
    "google-cloud-core==2.4.3",
    "google-cloud-storage==3.1.0",
    "google-crc32c==1.7.1",
    "google-resumable-media==2.7.2",
    "googleapis-common-protos==1.70.0",
    # greenlet 1.0.0 is broken on ppc64le
    # https://github.com/python-greenlet/greenlet/issues/230
    # by default program attempting to load an x86_64-only library from a native arm64 process
    # Beginning with gevent 20.12.0, 64-bit ARM binaries are distributed on PyPI for aarch64 manylinux2014
    # compatible systems. Resolves problem for m1 Mac chips
    "greenlet==3.0rc3",
    "gspread==3.6.0",
    "ibm-cos-sdk==2.14.1",
    "jinja2==3.1.6",
    "jsonschema>=3.2.0",
    "junitparser==3.1.0",
    "markdown==3.7",
    "keystoneauth1==5.11.0",  # Dependency of python-novaclient and python-cinderclient
    # new version of marshmallow 4.0.0 seems to be broken, failing with error:
    # TypeError: __init__() got an unexpected keyword argument 'default'
    "marshmallow==3.26.2",
    "msrestazure==0.6.3",
    "munch==2.5.0",
    "mysql-connector-python==9.1.0",
    "numpy==1.23.2",
    "oauth2client==4.1.3",
    "ocp-network-split>=0.3.1",
    "openshift==0.13.2",
    "openshift-python-wrapper==11.0.58",
    "oslo-i18n==6.5.1",  # Dependency of python-novaclient and python-cinderclient
    "oslo-serialization==5.7.0",  # Dependency of python-novaclient and python-cinderclient
    "oslo-utils==9.0.0",  # Dependency of python-novaclient and python-cinderclient
    "ovirt-engine-sdk-python==4.4.11",
    "pandas==1.5.2",
    "paramiko==3.5.0",
    "pexpect>=4.8.0",
    "ping3==4.0.3",
    "prettytable==0.7.2",
    "protobuf==4.25.8",
    "psutil==5.9.0",
    "psycopg2-binary==2.9.9",
    "pyopenssl==24.2.1",
    "pyotp==2.9.0",
    "pyparsing==2.4.7",
    "pytest==6.2.5",
    "pytest-html==3.1.1",
    "pytest-jira==0.3.22",
    "pytest-logger==0.5.1",
    "pytest-metadata==1.11.0",
    "pytest-order==1.2.0",
    "pytest-progress==1.2.5",
    "pytest-repeat==0.9.3",
    "python-cinderclient==9.7.0",
    "python-dateutil==2.9.0",
    "python-hcl2==3.0.1",
    "python-ipmi==0.5.7",
    "python-novaclient==18.9.0",
    "pyvmomi==7.0",
    "pyyaml==6.0.3",
    "range-key-dict==1.1.0",
    "reportportal-client==3.2.3",
    "requests==2.32.2",
    "scipy==1.12.0",
    "selenium==4.36.0",
    "semantic-version==2.8.5",
    "tabulate==0.9.0",
    "urllib3==2.6.3",
    "webdriver-manager==4.0.2",
    "xattr==1.3.0",
    "zope-event==6.0",
]

[tool.uv.sources]
ocp-network-split = { git = "https://github.com/red-hat-storage/ocp-network-split.git" }
detect-secrets = { git = "https://github.com/ibm/detect-secrets.git", rev = "master" }

[dependency-groups]
dev = [
    "black==24.3.0",
    "detect-secrets",
    "pre-commit==2.15.0",
    "tox==3.25.1",
]
docs = [
    "myst-parser>=4.0.1",
    "recommonmark==0.7.1",
    "sphinx==7.0.0",
    "sphinx-autopackagesummary==1.3",
    "sphinx-rtd-theme==2.0.0",
]

[project.scripts]
run-ci = "ocs_ci.framework.main:main"
run-ci-deploy = "ocs_ci.framework.deploy:main"
report-version = "ocs_ci.ocs.version:main"
ci-cleanup = "ocs_ci.cleanup.aws.cleanup:cluster_cleanup"
ci-pause = "ocs_ci.pause.pause:cluster_pause"
aws-cleanup = "ocs_ci.cleanup.aws.cleanup:aws_cleanup"
vsphere-cleanup = "ocs_ci.cleanup.vsphere.cleanup:vsphere_cleanup"
ocs-build = "ocs_ci.utility.ocs_build:main"
get-ssl-cert = "ocs_ci.utility.ssl_certs:main"
rosa-ocp-version = "ocs_ci.utility.rosa:rosa_ocp_version_endpoint"
deploy-fusion = "ocs_ci.framework.fusion.main:main"
deploy-fdf = "ocs_ci.framework.fusion_data_foundation.main:main"
analyze-logs = "ocs_ci.utility.log_analysis.cli:main"
analyze-trends = "ocs_ci.utility.log_analysis.cli:trends_main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
exclude = ["ez_setup"]
